Put the web api and caching process in an activity seperated from the main
