+++ MyApplication/app/src/main/java/com/github/ziazi/myapplication/MainActivity.java
@@ -1,6 +1,7 @@
 package com.github.ziazi.myapplication;
 
 import android.content.Context;
+import android.content.Intent;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Bundle;
@@ -20,73 +21,14 @@ import retrofit2.Response;
 
 public class MainActivity extends AppCompatActivity {
 
-    private  Callback<BoredActivity> callback_handler;
-    private ActivityDatabase db;
-
-    private final ExecutorService db_executor = Executors.newSingleThreadExecutor();
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        callback_handler = new Callback<BoredActivity>() {
-            @Override
-            public void onResponse(Call<BoredActivity> call, Response<BoredActivity> response) {
-                BoredActivity fetched= response.body();
-                if(fetched!=null) {
-                    fetched.link = "https://youtu.be/xvFZjo5PgG0";
-                    ((TextView) findViewById(R.id.textView)).setText(textFormat(fetched));
-                    ((TextView) findViewById(R.id.linkView)).setText(fetched.link);
-                    Linkify.addLinks((TextView) findViewById(R.id.linkView), Linkify.WEB_URLS);
-                    insertActivity(fetched);
-                }
-
-            }
-
-            @Override
-            public void onFailure(Call<BoredActivity> call, Throwable t) {
-                getCached();
-            }
-
-        };
-
-        this.db= Room.databaseBuilder(getApplicationContext(),
-                ActivityDatabase.class, "cache").build();
-
     }
 
-    public void getActivity(View view){
-        if(!isNetworkAvailable()) {
-            getCached();
-            return;
-        }
-        RetrofitClient.getClient().getApi().getActivity().enqueue(callback_handler);
-    }
-    private void insertActivity(BoredActivity activity){
-                db_executor.submit(()-> db.activityDao().insertAll(new CachedActivity().match(activity)));
-    }
-    private static String textFormat(BoredActivity fetched){
-       return  String.format("activity: %s \n accessibility: %s \n type: %s \n participants: %s \n price: %s \n key: %s \n ",
-                fetched.activity,fetched.accessibility,fetched.type,fetched.participants,fetched.price,fetched.key);
-    }
-
-    private void getCached(){
-        String error_message = "Offline Mode: Cannot connect to BoredApi";
-        ((TextView)findViewById(R.id.linkView)).setText(error_message);
-
-        db_executor.submit(()-> {
-                    CachedActivity loaded = db.activityDao().randomSelect();
-                    if(loaded!= null) {
-                        ((TextView) findViewById(R.id.textView)).setText(textFormat(loaded.revert()));
-                    }
-                }
-        );
-    }
-    private boolean isNetworkAvailable() {
-        ConnectivityManager connectivityManager
-                = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
-        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
-
+    public void fetchActivity(View view){
+        Intent intent = new Intent(this, FetchDataActivity.class);
+        startActivity(intent);
     }
 }
