+++ MyApplication/app/src/main/res/layout/activity_fetch_data.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".FetchDataActivity">
+
+    <Button
+        android:id="@+id/fetch"
+        android:layout_width="194dp"
+        android:layout_height="43dp"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="44dp"
+        android:onClick="getActivity"
+        android:text="Request Activity"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/mode" />
+
+    <TextView
+        android:id="@+id/activityText"
+        android:layout_width="311dp"
+        android:layout_height="150dp"
+        android:layout_marginStart="52dp"
+        android:layout_marginTop="212dp"
+        android:text="This contains the activity, press 'Request Activity' to get one :D"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/mode"
+        android:layout_width="245dp"
+        android:layout_height="68dp"
+        android:layout_marginStart="84dp"
+        android:layout_marginTop="96dp"
+        android:text="This is the mode: online/offline  :D"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/activityText" />
+</androidx.constraintlayout.widget.ConstraintLayout>
