+++ MyApplication/app/src/main/res/layout/activity_main.xml
@@ -7,38 +7,26 @@
     tools:context=".MainActivity"
     >
 
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="311dp"
-        android:layout_height="290dp"
-        android:layout_marginStart="96dp"
-        android:layout_marginTop="156dp"
-        android:text="Click on the Button to get an activity"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <Button
         android:id="@+id/button"
-        android:layout_width="112dp"
-        android:layout_height="88dp"
-        android:layout_marginStart="148dp"
-        android:layout_marginTop="96dp"
-        android:onClick="getActivity"
-        android:text="Suggest Activity"
+        android:layout_width="136dp"
+        android:layout_height="89dp"
+        android:layout_marginStart="136dp"
+        android:layout_marginBottom="100dp"
+        android:onClick="fetchActivity"
+        android:text="start"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView"
-        app:layout_constraintVertical_bias="0.0" />
+        app:layout_constraintTop_toBottomOf="@+id/greeting" />
 
     <TextView
-        android:id="@+id/linkView"
-        android:layout_width="275dp"
-        android:layout_height="17dp"
-        android:layout_marginTop="292dp"
-        android:layout_marginEnd="28dp"
-        android:text="Link of activity"
-        app:layout_constraintEnd_toEndOf="@+id/textView"
-        app:layout_constraintHorizontal_bias="0.923"
+        android:id="@+id/greeting"
+        android:layout_width="266dp"
+        android:layout_height="146dp"
+        android:layout_marginStart="72dp"
+        android:layout_marginTop="184dp"
+        android:text="Hello! Please click on the button to get an activity from BoredApi :D "
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
